feat(shield): add shield reduce damage logic
feat(shield): add shield reduce damage logic #2

feat(shield): rework reduced shield to percent logic #3

+++ Double Space Fighter/Assets/Scripts/Models/Stats/IUpdatableStat.cs
@@ -1,4 +1,4 @@
 ﻿public interface IUpdatableStat : IStat
 {
-    void Update(double dt);
+    void Update(float dt);
 }
